refactor retrievers
